Significant code cleanup, small analytics improvements, trying to lazy load model to reduce cold start time

+++ knn-handler/handler.py
@@ -1,5 +1,6 @@
 import base64
 import collections
+import functools
 import io
 import math
 import os
@@ -12,11 +13,6 @@ from google.cloud import storage, pubsub
 import numpy as np
 from PIL import Image
 import torch
-import torch.nn.functional as F
-
-from detectron2.layers import ShapeSpec
-from detectron2.checkpoint.detection_checkpoint import DetectionCheckpointer
-from detectron2.modeling.backbone.resnet import build_resnet_backbone
 
 import config
 
@@ -24,28 +20,49 @@ import config
 # Set up Google Cloud clients
 storage_client = storage.Client()
 publish_client = pubsub.PublisherClient()
-
-# Download model weights
 storage_bucket = storage_client.bucket(config.CLOUD_STORAGE_BUCKET)
-pathlib.Path(config.MODEL_LOCAL_PATH).parent.mkdir(parents=True, exist_ok=True)
-model_blob = storage_bucket.blob(config.MODEL_CLOUD_PATH)
-model_blob.download_to_filename(config.MODEL_LOCAL_PATH)
 
-# Create model
-shape = ShapeSpec(channels=3)
-model = torch.nn.Sequential(build_resnet_backbone(config.RESNET_CONFIG, shape))
 
-# Load model weights
-checkpointer = DetectionCheckpointer(model, save_to_disk=False)
-checkpointer.load(config.MODEL_LOCAL_PATH)
-model.eval()
+# Lazy load model
+model = None
 
-# Delete weights file to free up memory
-os.remove(config.MODEL_LOCAL_PATH)
 
-# Create global tensors for inference
-pixel_mean = torch.Tensor(config.RESNET_CONFIG.MODEL.PIXEL_MEAN).view(1, -1, 1, 1)
-pixel_std = torch.Tensor(config.RESNET_CONFIG.MODEL.PIXEL_STD).view(1, -1, 1, 1)
+def ensure_globals(f):
+    @functools.wraps(f)
+    def wrapper(*args, **kwargs):
+        kwargs["start_time"] = time.time()
+
+        global model
+        if not model:
+            from detectron2.layers import ShapeSpec
+            from detectron2.checkpoint.detection_checkpoint import DetectionCheckpointer
+            from detectron2.modeling.backbone.resnet import build_resnet_backbone
+
+            # Download model weights
+            pathlib.Path(config.MODEL_LOCAL_PATH).parent.mkdir(
+                parents=True, exist_ok=True
+            )
+            model_blob = storage_bucket.blob(config.MODEL_CLOUD_PATH)
+            model_blob.download_to_filename(config.MODEL_LOCAL_PATH)
+
+            # Create model
+            shape = ShapeSpec(channels=3)
+            model = torch.nn.Sequential(
+                build_resnet_backbone(config.get_resnet_config(), shape)
+            )
+
+            # Load model weights
+            checkpointer = DetectionCheckpointer(model, save_to_disk=False)
+            checkpointer.load(config.MODEL_LOCAL_PATH)
+            model.eval()
+
+            # Delete weights file to free up memory
+            os.remove(config.MODEL_LOCAL_PATH)
+
+        return f(*args, **kwargs)
+
+    return wrapper
+
 
 # Start web server
 worker_id = uuid.uuid4()
@@ -53,15 +70,18 @@ app = Flask(__name__)
 
 
 @app.route("/", methods=["POST"])
-def handler():
-    start_time = time.time()
+@ensure_globals
+def handler(start_time):
+    request_start_time = time.time()
 
     event = request.get_json()
     template = deserialize(event["template"]).unsqueeze(dim=0)
-    results = collections.defaultdict(dict)
 
-    results["metadata"]["worker_id"] = worker_id
-    results["metadata"]["compute"] = 0
+    results = {
+        "worker_id": worker_id,
+        "compute_time": 0.0,
+        "images": collections.defaultdict(dict),
+    }
 
     for image_cloud_path in event["images"]:
         try:
@@ -70,25 +90,24 @@ def handler():
             compute_start_time = time.time()
             result = run_inference(image)
             score_map, score = compute_knn_score(result, template)
-            results["metadata"]["compute"] += time.time() - compute_start_time
+            results["compute_time"] += time.time() - compute_start_time
 
-            results[image_cloud_path]["score"] = score
+            results["images"][image_cloud_path]["score"] = score
             if event.get("include_score_map"):
                 score_map_base64 = float_map_to_base64_png(score_map)
-                results[image_cloud_path]["score_map"] = score_map_base64
+                results["images"][image_cloud_path]["score_map"] = score_map_base64
         except AssertionError:
             pass
 
-    results["metadata"]["total"] = time.time() - start_time
+    results["request_time"] = time.time() - request_start_time
+    results["gcr_time"] = time.time() - start_time
 
     return jsonify(results)
 
-    # publish_client.publish(event["id"], json.dumps(results))
-    # return ""
-
 
 @app.route("/get_embedding", methods=["POST"])
-def get_embedding():
+@ensure_globals
+def get_embedding(start_time):
     event = request.get_json()
 
     image = download_image(event["image"])
@@ -99,7 +118,7 @@ def get_embedding():
     x2, y2 = [math.ceil(dim / config.RESNET_DOWNSAMPLE_FACTOR) for dim in (x2, y2)]
 
     embedding = result[:, y1:y2, x1:x2].mean(dim=-1).mean(dim=-1)
-    embedding = F.normalize(embedding, p=2, dim=0)
+    embedding = torch.nn.functional.normalize(embedding, p=2, dim=0)
     return serialize(embedding)
 
 
@@ -129,10 +148,17 @@ def image_to_tensor(image):
         image = torch.as_tensor(image).permute(2, 0, 1).unsqueeze(dim=0)
 
         # RGB -> BGR
-        if config.RESNET_CONFIG.INPUT.FORMAT == "BGR":
+        if config.get_resnet_config().INPUT.FORMAT == "BGR":
             image = torch.flip(image, dims=(1,))
-
         image = image.contiguous()
+
+        # Normalize
+        pixel_mean = torch.Tensor(config.get_resnet_config().MODEL.PIXEL_MEAN).view(
+            1, -1, 1, 1
+        )
+        pixel_std = torch.Tensor(config.get_resnet_config().MODEL.PIXEL_STD).view(
+            1, -1, 1, 1
+        )
         return (image - pixel_mean) / pixel_std
 
 
@@ -142,7 +168,7 @@ def compute_knn_score(embeddings, template):
 
     with torch.no_grad():
         embeddings_flat = embeddings.view(embeddings.size(0), -1)
-        embeddings_flat = F.normalize(embeddings_flat, p=2, dim=0)
+        embeddings_flat = torch.nn.functional.normalize(embeddings_flat, p=2, dim=0)
         scores = cosine_similarity(template, embeddings_flat).view(-1)
         score_map = scores.view(embeddings.size(1), embeddings.size(2))
         score = torch.topk(scores, config.N_DISTANCES_TO_AVERAGE).values.mean().item()
